requirements.txt
#python  v3.12.2 
colorlog
requests
python-dotenv
sqlite3
os
numpy
pandas


